+++ tests/test_search.py
@@ -3,12 +3,12 @@
 
 from typing import Any
 
-from async_upnp_client.search import SSDPListener
+from async_upnp_client.search import SsdpSearchListener
 from async_upnp_client.ssdp import SSDP_IP_V4
 
 
 def test_create_ssdp_listener_with_alternate_target() -> None:
-    """Create a SSDPListener on an alternate target."""
+    """Create a SsdpSearchListener on an alternate target."""
 
     async def _dummy_callback(*_: Any) -> None:
         # Nop.
@@ -16,7 +16,7 @@ def test_create_ssdp_listener_with_alternate_target() -> None:
 
     yeelight_target = (SSDP_IP_V4, 1982)
     yeelight_service_type = "wifi_bulb"
-    listener = SSDPListener(
+    listener = SsdpSearchListener(
         async_callback=_dummy_callback,
         async_connect_callback=_dummy_callback,
         service_type=yeelight_service_type,